+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/raw/RawService.kt
@@ -19,7 +19,7 @@ package org.matrix.android.sdk.api.raw
 import org.matrix.android.sdk.api.cache.CacheStrategy
 
 /**
- * Useful methods to fetch raw data from the server. The access token will not be used to fetched the data
+ * Useful methods to fetch raw data from the server. The access token will not be used to fetch the data
  */
 interface RawService {
     /**